# documentation/docs/content/openAPI/spring_backend-openapi.yaml
openapi: "3.0.3"
info:
  title: "spring_backend API"
  description: "spring_backend API"
  version: "1.0.0"
servers:
  - url: "https://spring_backend"
paths:
...
  /api/v1/auth/signup:
    post:
      summary: "POST api/v1/auth/signup"
      operationId: "signup"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignUpRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JwtAuthenticationResponse"
  /api/v1/auth/login:
    post:
      summary: "POST api/v1/auth/login"
      operationId: "login"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SigninRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JwtAuthenticationResponse"
  /api/v1/auth/refresh:
    post:
      summary: "POST api/v1/auth/refresh"
      operationId: "refresh"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RefreshTokenResponse"
...
components:
  schemas:
...
    SignUpRequest:
      type: "object"
      properties:
        firstName:
          type: "string"
        lastName:
          type: "string"
        email:
          type: "string"
        password:
          type: "string"
        languageCode:
          type: "string"
    JwtAuthenticationResponse:
      type: "object"
      properties:
        token:
          type: "string"
        refreshToken:
          type: "string"
        refreshTokenTTL:
          type: "string"
        languageCode:
          type: "string"
    SigninRequest:
      type: "object"
      properties:
        email:
          type: "string"
        password:
          type: "string"
    RefreshTokenRequest:
      type: "object"
      properties:
        token:
          type: "string"
    RefreshTokenResponse:
      type: "object"
      properties:
        token:
          type: "string"
        refreshToken:
          type: "string"
...